backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-community
langchain-core
langchain-openai
langchain-google-genai
chromadb
python-dotenv
sqlmodel
pydantic
pydantic-settings
google-generativeai